add: use collection volume
